requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
langchain[google-genai]
langchain-community
langchain-google-community
PyPDF2
docx2txt
pydub
noisereduce
soundfile
librosa
moviepy
speechrecognition
python-dotenv
jina-embeddings-v2